Logical Not: Moves tests to logical ops tests, checks for dtype to be bool
