+++ src/math/logicalop_test.ts
@@ -21,12 +21,81 @@ import * as util from '../util';
 
 import {Array1D, Array2D, Array3D, Array4D} from './ndarray';
 
+// LogicalNot:
+{
+  const boolNaN = util.getNaN('bool');
+
+  const tests: MathTests = it => {
+    it('Array1D.', math => {
+      let a = Array1D.new([1, 0, 0], 'bool');
+      test_util.expectArraysClose(math.logicalNot(a), [0, 1, 1]);
+
+      a = Array1D.new([0, 0, 0], 'bool');
+      test_util.expectArraysClose(math.logicalNot(a), [1, 1, 1]);
+
+      a = Array1D.new([1, 1], 'bool');
+      test_util.expectArraysClose(math.logicalNot(a), [0, 0]);
+    });
+    it('NaNs in Array1D', math => {
+      const a = Array1D.new([1, NaN, 0], 'bool');
+      test_util.expectArraysClose(math.logicalNot(a), [0, boolNaN, 1]);
+    });
+
+    it('Array2D', math => {
+      let a = Array2D.new([2, 3], [[1, 0, 1], [0, 0, 0]], 'bool');
+      test_util.expectArraysClose(math.logicalNot(a), [0, 1, 0, 1, 1, 1]);
+
+      a = Array2D.new([2, 3], [[0, 0, 0], [1, 1, 1]], 'bool');
+      test_util.expectArraysClose(math.logicalNot(a), [1, 1, 1, 0, 0, 0]);
+    });
+    it('NaNs in Array2D', math => {
+      const a = Array2D.new([2, 2], [[1, NaN], [0, NaN]], 'bool');
+      test_util.expectArraysClose(math.logicalNot(a), [0, boolNaN, 1, boolNaN]);
+    });
+
+    it('Array3D', math => {
+      let a =
+          Array3D.new([2, 3, 1], [[[1], [0], [1]], [[0], [0], [0]]], 'bool');
+      test_util.expectArraysClose(math.logicalNot(a), [0, 1, 0, 1, 1, 1]);
+
+      a = Array3D.new([2, 3, 1], [[[0], [0], [0]], [[1], [1], [1]]], 'bool');
+      test_util.expectArraysClose(math.logicalNot(a), [1, 1, 1, 0, 0, 0]);
+    });
+    it('NaNs in Array3D', math => {
+      const a =
+          Array3D.new([2, 3, 1], [[[1], [NaN], [1]], [[0], [0], [0]]], 'bool');
+      test_util.expectArraysClose(math.logicalNot(a), [0, boolNaN, 0, 1, 1, 1]);
+    });
+
+    it('Array4D', math => {
+      let a = Array4D.new([2, 2, 1, 1], [1, 0, 1, 0], 'bool');
+      test_util.expectArraysClose(math.logicalNot(a), [0, 1, 0, 1]);
+
+      a = Array4D.new([2, 2, 1, 1], [0, 0, 0, 0], 'bool');
+      test_util.expectArraysClose(math.logicalNot(a), [1, 1, 1, 1]);
+
+      a = Array4D.new([2, 2, 1, 1], [1, 1, 1, 1], 'bool');
+      test_util.expectArraysClose(math.logicalNot(a), [0, 0, 0, 0]);
+    });
+    it('NaNs in Array4D', math => {
+      const a = Array4D.new([2, 2, 1, 1], [1, NaN, 1, 0], 'bool');
+      test_util.expectArraysClose(math.logicalNot(a), [0, boolNaN, 0, 1]);
+    });
+  };
+
+  test_util.describeMathCPU('logicalNot', [tests]);
+  test_util.describeMathGPU('logicalNot', [tests], [
+    {'WEBGL_FLOAT_TEXTURE_ENABLED': true, 'WEBGL_VERSION': 1},
+    {'WEBGL_FLOAT_TEXTURE_ENABLED': true, 'WEBGL_VERSION': 2},
+    {'WEBGL_FLOAT_TEXTURE_ENABLED': false, 'WEBGL_VERSION': 1}
+  ]);
+}
+
 // LogicalAnd:
 {
   const boolNaN = util.getNaN('bool');
 
   const tests: MathTests = it => {
-    // Array1D:
     it('Array1D.', math => {
       let a = Array1D.new([1, 0, 0], 'bool');
       let b = Array1D.new([0, 1, 0], 'bool');
@@ -54,7 +123,6 @@ import {Array1D, Array2D, Array3D, Array4D} from './ndarray';
       test_util.expectArraysClose(math.logicalAnd(a, b), [0, boolNaN, boolNaN]);
     });
 
-    // Array2D:
     it('Array2D', math => {
       let a = Array2D.new([2, 3], [[1, 0, 1], [0, 0, 0]], 'bool');
       let b = Array2D.new([2, 3], [[0, 0, 0], [0, 1, 0]], 'bool');
@@ -76,7 +144,6 @@ import {Array1D, Array2D, Array3D, Array4D} from './ndarray';
           math.logicalAnd(a, b), [0, boolNaN, 0, boolNaN]);
     });
 
-    // Array3D:
     it('Array3D', math => {
       let a =
           Array3D.new([2, 3, 1], [[[1], [0], [1]], [[0], [0], [1]]], 'bool');
@@ -106,7 +173,6 @@ import {Array1D, Array2D, Array3D, Array4D} from './ndarray';
           math.logicalAnd(a, b), [0, boolNaN, 1, 0, 0, boolNaN]);
     });
 
-    // Array4D:
     it('Array4D', math => {
       let a = Array4D.new([2, 2, 1, 1], [1, 0, 1, 0], 'bool');
       let b = Array4D.new([2, 2, 1, 1], [0, 1, 1, 0], 'bool');
@@ -148,7 +214,6 @@ import {Array1D, Array2D, Array3D, Array4D} from './ndarray';
   const boolNaN = util.getNaN('bool');
 
   const tests: MathTests = it => {
-    // Array1D:
     it('Array1D.', math => {
       let a = Array1D.new([1, 0, 0], 'bool');
       let b = Array1D.new([0, 1, 0], 'bool');
@@ -176,7 +241,6 @@ import {Array1D, Array2D, Array3D, Array4D} from './ndarray';
       test_util.expectArraysClose(math.logicalOr(a, b), [1, boolNaN, boolNaN]);
     });
 
-    // Array2D:
     it('Array2D', math => {
       let a = Array2D.new([2, 3], [[1, 0, 1], [0, 0, 0]], 'bool');
       let b = Array2D.new([2, 3], [[0, 0, 0], [0, 1, 0]], 'bool');
@@ -198,7 +262,6 @@ import {Array1D, Array2D, Array3D, Array4D} from './ndarray';
           math.logicalOr(a, b), [1, boolNaN, 1, boolNaN]);
     });
 
-    // Array3D:
     it('Array3D', math => {
       let a =
           Array3D.new([2, 3, 1], [[[1], [0], [1]], [[0], [0], [0]]], 'bool');
@@ -228,7 +291,6 @@ import {Array1D, Array2D, Array3D, Array4D} from './ndarray';
           math.logicalOr(a, b), [1, boolNaN, 1, 1, 0, boolNaN]);
     });
 
-    // Array4D:
     it('Array4D', math => {
       let a = Array4D.new([2, 2, 1, 1], [1, 0, 1, 0], 'bool');
       let b = Array4D.new([2, 2, 1, 1], [0, 1, 0, 0], 'bool');
